Add original code written by @jansuategui
Credit goes to @jansuategui for writing this code. Code was provided to me by @cleme.

Tested it against a couple of datasets and it appears to be working with biom-format 1.2.0, PyCogent 1.5.3, numpy 1.5.1, and QIIME 1.7.0-dev, master@36db13d. Unit tests pass after fixing an import error.

Removed csmat.py as it can just be imported from the biom-format package.